Move `Swarm.Game.Value` -> `Swarm.Language.Value`
It's always bothered me that `Value` lived under the `Game` namespace;
I don't know why I did it that way initially.  It really belongs with
`Language`, since the type of values that result from
evaluating/executing a program are part of the definition of a
language.
