+++ test/unit/TestEval.hs
@@ -9,7 +9,7 @@ import Data.Char (ord)
 import Data.Text (Text)
 import Data.Text qualified as T
 import Swarm.Game.State
-import Swarm.Game.Value
+import Swarm.Language.Value
 import Test.Tasty
 import Test.Tasty.HUnit
 import Test.Tasty.QuickCheck
